requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.9.5
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
asttokens==3.0.0
attrs==24.3.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
dataclasses-json==0.6.7
debugpy==1.8.11
decorator==5.1.1
distro==1.9.0
executing==2.1.0
fastapi==0.115.6
frozenlist==1.5.0
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-core==2.4.1
google-cloud-documentai==3.1.0
google-cloud-translate==3.19.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
groq==0.15.0
grpc-google-iam-v1==0.14.0
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.5
jedi==0.19.2
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.29
langchain-experimental==0.3.4
langchain-groq==0.2.3
langchain-neo4j==0.2.0
langchain-openai==0.3.0
langchain-pinecone==0.2.1
langchain-tests==0.3.8
langchain-text-splitters==0.3.5
langsmith==0.2.10
marshmallow==3.25.0
matplotlib-inline==0.1.7
multidict==6.1.0
mypy-extensions==1.0.0
neo4j==5.27.0
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.59.6
orjson==3.10.14
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pinecone==5.4.2
pinecone-plugin-inference==3.1.0
pinecone-plugin-interface==0.0.7
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.48
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
pypdf==5.1.0
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-socket==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
stack-data==0.6.3
starlette==0.41.3
syrupy==4.8.0
tenacity==9.0.0
tiktoken==0.8.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
widgetsnbextension==4.0.13
yarl==1.18.3
yfiles_jupyter_graphs==1.10.1